// interfaces/Filters.tsx
export default interface Filters {
  city: string;
  country: string;
  startDate: string;
  endDate: string;
  adults: number;
  children: number;
  pets: number;
  typeOfPlace: number;
  minPrice: number;
  maxPrice: number;
  bedrooms: number;
  beds: number;
  bathrooms: number;
  rating: number;
  accomodationType: number;
  amenities: number[];
}